3619: Rename EmptyAttributeNameIssueGenerator files
