+++ common/src/View/MapDocument.cpp
@@ -45,10 +45,10 @@
 #include "Model/BrushGeometry.h"
 #include "Model/ChangeBrushFaceAttributesRequest.h"
 #include "Model/EditorContext.h"
-#include "Model/EmptyAttributeNameIssueGenerator.h"
 #include "Model/EmptyAttributeValueIssueGenerator.h"
 #include "Model/EmptyBrushEntityIssueGenerator.h"
 #include "Model/EmptyGroupIssueGenerator.h"
+#include "Model/EmptyPropertyKeyIssueGenerator.h"
 #include "Model/Entity.h"
 #include "Model/EntityNode.h"
 #include "Model/Game.h"
